% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsc.R
\name{cpsc_recall}
\alias{cpsc_recall}
\title{Get a single CPSC recall by recall number}
\usage{
cpsc_recall(recall_number)
}
\arguments{
\item{recall_number}{Recall number (e.g. "24090").}
}
\value{
tibble with one row of recall details
}
\description{
Get a single CPSC recall by recall number
}
